Add project structure and files
The project structure for this project was generated by Astro.

All the necessary assets for the project were added to the public folder
, this for have all this assets in a public folder just as Astro do with
its project structure, also the input.css was added.

And finally in the tailwind.config.mjs were added the fonts and colors
to be used in the project.